# specs/openapi.yaml
openapi: 3.0.0
info:
  title: Restaurant Ordering API
  description: A restaurant ordering system with authentication and restaurant management
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.restaurant.example.com/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /restaurants:
    get:
      summary: List all restaurants
      description: Retrieves a paginated list of all active restaurants. Supports pagination via page and size parameters.
      operationId: getRestaurants
      security: []
      parameters:
        - name: page
          in: query
          description: Page number (0-indexed) for pagination
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 10000
        - name: size
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantPage"
        "400":
          description: Invalid query parameters
        "500":
          description: Internal server error
    post:
      summary: Create a new restaurant
      description: Creates a new restaurant entry. Requires ADMIN role. The authenticated user will be set as the restaurant owner.
      operationId: createRestaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantRequest"
      responses:
        "201":
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized - invalid or missing token
        "403":
          description: Forbidden - requires admin role
  /restaurants/{restaurantId}:
    get:
      summary: Get a specific restaurant
      description: Retrieves detailed information about a specific restaurant by its ID.
      operationId: getRestaurant
      security: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          description: Unique identifier of the restaurant
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Restaurant found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "400":
          description: Invalid restaurantId format
        "404":
          description: Restaurant not found
        "500":
          description: Internal server error
    put:
      summary: Update a restaurant
      description: Updates an existing restaurant's information. Requires ADMIN role.
      operationId: updateRestaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          description: Unique identifier of the restaurant to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRestaurantRequest"
      responses:
        "200":
          description: Restaurant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "400":
          description: Invalid restaurantId format or request body
        "401":
          description: Unauthorized - invalid or missing token
        "403":
          description: Forbidden - requires admin role
        "404":
          description: Restaurant not found
    delete:
      summary: Delete a restaurant
      description: Permanently deletes a restaurant and all associated data. Requires ADMIN role.
      operationId: deleteRestaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          description: Unique identifier of the restaurant to delete
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Restaurant deleted
        "400":
          description: Invalid restaurantId format
        "401":
          description: Unauthorized - invalid or missing token
        "403":
          description: Forbidden - requires admin role
        "404":
          description: Restaurant not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for Spring Security authentication
  schemas:
...
    Restaurant:
      type: object
      description: Restaurant business information
      properties:
        id:
          type: string
          format: uuid
          description: Unique restaurant identifier (UUID)
        name:
          type: string
          description: Restaurant business name
        description:
          type: string
          description: Brief description of the restaurant
        address:
          type: string
          description: Physical address of the restaurant
        phone:
          type: string
          description: Contact phone number
        email:
          type: string
          format: email
          description: Contact email address
        cuisineType:
          type: string
          description: Type of cuisine (e.g., Italian, Asian, American)
        openingHours:
          type: string
          description: "Business hours (e.g., Mon-Fri: 9AM-10PM)"
        ownerId:
          type: string
          format: uuid
          description: User ID of the restaurant owner
        isActive:
          type: boolean
          description: Whether the restaurant is currently active and accepting orders
        createdAt:
          type: string
          format: date-time
          description: Restaurant creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    CreateRestaurantRequest:
      type: object
      description: Request body for creating a new restaurant
      additionalProperties: false
      required:
        - name
        - address
      properties:
        name:
          type: string
          description: Restaurant business name
          minLength: 1
          maxLength: 100
        description:
          type: string
          description: Brief description of the restaurant (optional)
          maxLength: 500
        address:
          type: string
          description: Physical address of the restaurant
          minLength: 1
          maxLength: 200
        phone:
          type: string
          description: Contact phone number (optional)
          maxLength: 20
        email:
          type: string
          format: email
          description: Contact email address (optional)
        cuisineType:
          type: string
          description: Type of cuisine served (optional)
          maxLength: 50
        openingHours:
          type: string
          description: Operating hours (optional)
          maxLength: 100
    UpdateRestaurantRequest:
      type: object
      description: Request body for updating an existing restaurant
      additionalProperties: false
      properties:
        name:
          type: string
          description: Restaurant business name (optional)
          minLength: 1
          maxLength: 100
        description:
          type: string
          description: Brief description of the restaurant (optional)
          maxLength: 500
        address:
          type: string
          description: Physical address of the restaurant (optional)
          maxLength: 200
        phone:
          type: string
          description: Contact phone number (optional)
          maxLength: 20
        email:
          type: string
          format: email
          description: Contact email address (optional)
        cuisineType:
          type: string
          description: Type of cuisine served (optional)
          maxLength: 50
        openingHours:
          type: string
          description: Operating hours (optional)
          maxLength: 100
        isActive:
          type: boolean
          description: Whether the restaurant is active (optional)
    RestaurantPage:
      type: object
      description: Paginated response for restaurant listings
      properties:
        content:
          type: array
          description: List of restaurants on this page
          items:
            $ref: "#/components/schemas/Restaurant"
        totalElements:
          type: integer
          description: Total number of restaurants available
        totalPages:
          type: integer
          description: Total number of pages
        number:
          type: integer
          description: Current page number (0-indexed)
        size:
          type: integer
          description: Number of items per page
...